test: LottoPrintTest 로또 계산기 테스트와 분리 및 전체 테스트 통과
