+++ __tests__/LottoCalculatorTest.js
@@ -1,12 +1,5 @@
 const LottoCalculate = require("../src/LottoCalculator");
-const MissionUtils = require("@woowacourse/mission-utils");
 const Lotto = require("../src/Lotto");
-
-const getLogSpy = () => {
-  const logSpy = jest.spyOn(MissionUtils.Console, "print");
-  logSpy.mockClear();
-  return logSpy;
-};
 
 const winLotto = [1, 2, 3, 4, 5, 6];
 const bonusNum = 8;
@@ -34,10 +27,13 @@ describe("로또 당첨 결과 테스트", () => {
     return new Lotto(lotto);
   });
 
-  const result = lottoCalculate.resultCaculator(lottoArr, {
-    winLotto: new Lotto(winLotto),
-    bonus: bonusNum,
-  });
+  const [result, gain] = lottoCalculate
+    .resultCaculator(lottoArr, {
+      winLotto: new Lotto(winLotto),
+      bonus: bonusNum,
+    })
+    .gainPercent()
+    .getLottoResult();
 
   test("로또 당첨 결과 확인 테스트 ", () => {
     expect(result).toEqual({
@@ -46,31 +42,7 @@ describe("로또 당첨 결과 테스트", () => {
       THIRD: 1,
       FOURTH: 1,
       FIFTH: 1,
-      count: 5,
     });
-  });
-
-  test("로또 당첨 결과 출력 테스트", () => {
-    const logs = [
-      "3개 일치 (5,000원) - 1개",
-      "4개 일치 (50,000원) - 1개",
-      "5개 일치 (1,500,000원) - 1개",
-      "5개 일치, 보너스 볼 일치 (30,000,000원) - 1개",
-      "6개 일치 (2,000,000,000원) - 1개",
-    ];
-
-    const logSpy = getLogSpy();
-    lottoCalculate.printWinResult(result);
-    logs.forEach((log) => {
-      expect(logSpy).toHaveBeenCalledWith(expect.stringContaining(log));
-    });
-  });
-
-  test("로또 수익률 출력 테스트", () => {
-    const logSpy = getLogSpy();
-    lottoCalculate.printGainPercent(result);
-    expect(logSpy).toHaveBeenCalledWith(
-      expect.stringContaining("총 수익률은 40631100.0%입니다.")
-    );
+    expect(gain).toEqual("40631100.0");
   });
 });
